refactored controller/model variable names
